apply loan application api - default values and not null rules added

+++ src/test/java/homework/repositories/LoanApplicationRepositoryIntegrationTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import homework.models.LoanApplication;
 import homework.models.LoanApplicationStatus;
 import net.bytebuddy.utility.RandomString;
-import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,7 +12,6 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.util.RandomAccess;
 
 
 @RunWith(SpringRunner.class)
@@ -66,7 +64,7 @@ public class LoanApplicationRepositoryIntegrationTest {
                 .email("mail@mail.lv")
                 .phone("324535")
                 .yearlyTurnover(100f)
-                .term((byte) 5)
+                .term((short) 5)
                 .status(LoanApplicationStatus.APPLIED).build();
     }
